*: move internal "etcdserver/api/rafthttp"
