+++ etcdserver/api/etcdhttp/peer_test.go
@@ -26,9 +26,9 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/coreos/etcd/etcdserver/api/membership"
+	"github.com/coreos/etcd/etcdserver/api/rafthttp"
 	"github.com/coreos/etcd/pkg/testutil"
 	"github.com/coreos/etcd/pkg/types"
-	"github.com/coreos/etcd/rafthttp"
 	"github.com/coreos/go-semver/semver"
 )
