+++ etcdserver/util.go
@@ -19,8 +19,8 @@ import (
 	"time"
 
 	"github.com/coreos/etcd/etcdserver/api/membership"
+	"github.com/coreos/etcd/etcdserver/api/rafthttp"
 	"github.com/coreos/etcd/pkg/types"
-	"github.com/coreos/etcd/rafthttp"
 
 	"go.uber.org/zap"
 )
